Setting up Project Controller, Service, Repository and Data Structure

+++ src/main/java/io/seunwater/openseun/requests/ListProjectRequest.java
@@ -19,7 +19,7 @@ public class ListProjectRequest {
 
     private String location;
 
-    private BigDecimal estimatedReductions;
+    private Double estimatedReductions;
 
     private ProjectCategory category;
